events: move CommonEvent out of bpf
Having it in its own file makes things cleaner.
