+++ src/events/mod.rs
@@ -21,6 +21,7 @@ pub(crate) use display::*;
 pub(crate) mod bpf;
 pub(crate) mod file;
 
+pub(crate) mod common;
 pub(crate) mod ct;
 pub(crate) mod kernel;
 pub(crate) mod nft;
