+++ src/process/series.rs
@@ -10,7 +10,7 @@ use std::collections::{BTreeMap, VecDeque};
 use anyhow::{anyhow, Result};
 
 use crate::{
-    events::{bpf::CommonEvent, skb_tracking::TrackingInfo, Event},
+    events::{common::CommonEvent, skb_tracking::TrackingInfo, Event},
     module::ModuleId,
 };
